chore: rename AssemblyScript SDK directory from `as` to `assemblyscript`
