+++ sdk/assemblyscript/http/cgi.ts
