+++ sdk/assemblyscript/http/http.ts
