+++ sdk/assemblyscript/index.ts
