+++ sdk/assemblyscript/package.json
